\name{read_twin_spec}
\alias{read_twin_spec}
\title{Read a twinfake sensitivity specification}
\usage{read_twin_spec(path)}
\arguments{\item{path}{Path to a JSON specification.}}
\value{A list with class \code{twinfake_spec}.}
\description{Reads a JSON specification created by \code{write_twin_spec()}.}
